[PageSpeed]: Removed Create and Configure folders, Polished Setup folder to ensure consistency with manual testing
